backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
aiofiles>=23.2.1
pydantic>=2.9.0
pdfplumber>=0.11.0
pypdf>=4.0.0
groq>=0.4.0
python-dotenv>=1.0.0
requests>=2.32.0
streamlit
pandas
requests
Pillow
pytest
psutil
reportlab

# Traditional / Core
PyMuPDF
pdfplumber
pdfminer.six

# OCR
pytesseract
easyocr
paddleocr
python-doctr

# Table Extraction
camelot-py
tabula-py
jpype1  # optional: tabula-py uses in-process Java; without it you see "Fallback to subprocess" (still works)

# Layout / Structure
layoutparser
opencv-python
unstructured[all-docs]

# Modern / LLM-based
docling
llmsherpa
marker-pdf
mineru
grobid-client-python
rapidocr
rapidocr-pdf
liteparse
# Surya OCR (suryaocr.py): transformers 5.x breaks ROPE_INIT_FUNCTIONS / Qwen2 decoder
transformers>=4.56.1,<5.0
surya-ocr

# Optional best-effort integrations (install manually if available):
# magic-pdf
